webcrawlerdemo/requirements.txt
attrs==19.3.0
Automat==0.8.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
dj-database-url==0.5.0
Django==2.0.12
django-crispy-forms==1.7.2
djangorestframework==3.9.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.6
incremental==17.5.0
lxml==4.4.2
parsel==1.5.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
python-decouple==3.3
python-dotenv==0.10.3
python-scrapyd-api==2.1.2
pytz==2019.3
queuelib==1.5.0
requests==2.18.4
Scrapy==1.5.1
scrapy-djangoitem==1.1.1
scrapyd==1.2.0
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.13.0
SQLAlchemy==1.3.11
Twisted==19.10.0
urllib3==1.22
w3lib==1.21.0
whitenoise==4.1.4
